refactor: move convert ckETH to ETH to a new top pkg
