+++ src/frontend/src/eth/components/send/SendNetwork.svelte
@@ -5,7 +5,7 @@
 	import { ETHEREUM_NETWORK, ICP_NETWORK } from '$lib/constants/networks.constants';
 	import { isEthAddress, isIcpAccountIdentifier } from '$lib/utils/account.utils';
 	import { isCkEthHelperContract } from '$eth/utils/send.utils';
-	import { ckEthHelperContractAddressStore } from '$eth/stores/cketh.store';
+	import { ckEthHelperContractAddressStore } from '$icp-eth/stores/cketh.store';
 	import { tokenId } from '$lib/derived/token.derived';
 
 	export let network: Network | undefined = undefined;
