+++ src/frontend/src/eth/services/fee.services.ts
@@ -3,9 +3,9 @@ import { ERC20_FALLBACK_FEE } from '$eth/constants/erc20.constants';
 import { ETH_BASE_FEE } from '$eth/constants/eth.constants';
 import { getFeeData as getBurnFeeData } from '$eth/providers/infura-erc20-icp.providers';
 import { getFeeData as getErc20FeeDataProvider } from '$eth/providers/infura-erc20.providers';
-import type { CkEthHelperContractAddressData } from '$eth/stores/cketh.store';
 import type { Erc20ContractAddress } from '$eth/types/erc20';
 import { isCkEthHelperContract } from '$eth/utils/send.utils';
+import type { CkEthHelperContractAddressData } from '$icp-eth/stores/cketh.store';
 import { ETHEREUM_NETWORK } from '$lib/constants/networks.constants';
 import type { ETH_ADDRESS } from '$lib/types/address';
 import type { Network } from '$lib/types/network';
