+++ svelte.config.js
@@ -27,7 +27,8 @@ const config = {
 		alias: {
 			$declarations: './src/declarations',
 			$eth: './src/frontend/src/eth',
-			$icp: './src/frontend/src/icp'
+			$icp: './src/frontend/src/icp',
+			'$icp-eth': './src/frontend/src/icp-eth'
 		}
 	},
